Move plus dialog to separate component
